+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
+### Added
+- Improved error handling for login
+
 ## [0.2.1] - 2021-05-28
 ### Fixed
 - Make weconnect object printable by __str__ method
@@ -17,7 +20,7 @@ Minor fix in observer interface
 ## [0.1.0] - 2021-05-26
 Initial release
 
-[unreleased]: https://github.com/tillsteinbach/WeConnect-python/compare/v0.2.0...HEAD
+[unreleased]: https://github.com/tillsteinbach/WeConnect-python/compare/v0.2.1...HEAD
 [0.2.1]: https://github.com/tillsteinbach/WeConnect-python/releases/tag/v0.2.1
 [0.2.0]: https://github.com/tillsteinbach/WeConnect-python/releases/tag/v0.2.0
 [0.1.1]: https://github.com/tillsteinbach/WeConnect-python/releases/tag/v0.1.1